sort out startup logging
