+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/config/Configuration.java
@@ -7,7 +7,6 @@ package io.kroxylicious.proxy.config;
 
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -26,14 +25,9 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 import io.kroxylicious.proxy.config.admin.AdminHttpConfiguration;
-import io.kroxylicious.proxy.config.tls.AllowDeny;
-import io.kroxylicious.proxy.config.tls.Tls;
-import io.kroxylicious.proxy.config.tls.TrustOptions;
-import io.kroxylicious.proxy.config.tls.TrustProvider;
 import io.kroxylicious.proxy.model.VirtualClusterModel;
 import io.kroxylicious.proxy.service.ClusterNetworkAddressConfigProvider;
 import io.kroxylicious.proxy.service.ClusterNetworkAddressConfigProviderService;
-import io.kroxylicious.proxy.service.HostPort;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
@@ -184,6 +178,7 @@ public record Configuration(
                 .filter(Predicate.not(List::isEmpty))
                 .ifPresentOrElse(listeners -> addListeners(pfr, listeners, virtualClusterModel),
                         () -> addListenerFromDeprecatedConfig(virtualCluster, pfr, virtualClusterModel));
+        virtualClusterModel.logVirtualClusterSummary();
 
         return virtualClusterModel;
     }
@@ -193,8 +188,6 @@ public record Configuration(
             var networkAddress = buildNetworkAddressProviderService(listener.clusterNetworkAddressConfigProvider(), pfr);
             var tls = listener.tls();
             virtualClusterModel.addListener(listener.name(), networkAddress, tls);
-            logVirtualClusterSummary(virtualClusterModel.getClusterName(), virtualClusterModel.targetCluster(), networkAddress,
-                    tls);
         });
     }
 
@@ -205,47 +198,6 @@ public record Configuration(
         Objects.requireNonNull(virtualCluster.clusterNetworkAddressConfigProvider(), "provider unexpectedly null");
         var networkAddress = buildNetworkAddressProviderService(virtualCluster.clusterNetworkAddressConfigProvider(), pfr);
         virtualClusterModel.addListener(VirtualCluster.DEFAULT_LISTENER_NAME, networkAddress, virtualCluster.tls());
-    }
-
-    @SuppressWarnings("java:S1874") // the classes are deprecated because we don't want them in the API module
-    private static void logVirtualClusterSummary(String clusterName,
-                                                 TargetCluster targetCluster,
-                                                 ClusterNetworkAddressConfigProvider clusterNetworkAddressConfigProvider,
-                                                 Optional<Tls> tls) {
-        try {
-            HostPort downstreamBootstrap = clusterNetworkAddressConfigProvider.getClusterBootstrapAddress();
-            var downstreamTlsSummary = generateTlsSummary(tls);
-
-            HostPort upstreamHostPort = targetCluster.bootstrapServersList().get(0);
-            var upstreamTlsSummary = generateTlsSummary(targetCluster.tls());
-
-            LOGGER.info("Virtual Cluster: {}, Downstream {}{} => Upstream {}{}",
-                    clusterName, downstreamBootstrap, downstreamTlsSummary, upstreamHostPort, upstreamTlsSummary);
-        }
-        catch (Exception e) {
-            LOGGER.warn("Failed to log summary for Virtual Cluster: {}", clusterName, e);
-        }
-    }
-
-    private static String generateTlsSummary(Optional<Tls> tlsToSummarize) {
-        var tls = tlsToSummarize.map(t -> Optional.ofNullable(t.trust())
-                .map(TrustProvider::trustOptions)
-                .map(TrustOptions::toString).orElse("-"))
-                .map(options -> " (TLS: " + options + ") ").orElse("");
-        var cipherSuitesAllowed = tlsToSummarize.map(t -> Optional.ofNullable(t.cipherSuites())
-                .map(AllowDeny::allowed).orElse(Collections.emptyList()))
-                .map(allowedCiphers -> " (Allowed Ciphers: " + allowedCiphers + ")").orElse("");
-        var cipherSuitesDenied = tlsToSummarize.map(t -> Optional.ofNullable(t.cipherSuites())
-                .map(AllowDeny::denied).orElse(Collections.emptySet()))
-                .map(deniedCiphers -> " (Denied Ciphers: " + deniedCiphers + ")").orElse("");
-        var protocolsAllowed = tlsToSummarize.map(t -> Optional.ofNullable(t.protocols())
-                .map(AllowDeny::allowed).orElse(Collections.emptyList()))
-                .map(protocols -> " (Allowed Protocols: " + protocols + ")").orElse("");
-        var protocolsDenied = tlsToSummarize.map(t -> Optional.ofNullable(t.protocols())
-                .map(AllowDeny::denied).orElse(Collections.emptySet()))
-                .map(protocols -> " (Denied Protocols: " + protocols + ")").orElse("");
-
-        return tls + cipherSuitesAllowed + cipherSuitesDenied + protocolsAllowed + protocolsDenied;
     }
 
     private static ClusterNetworkAddressConfigProvider buildNetworkAddressProviderService(@NonNull ClusterNetworkAddressConfigProviderDefinition definition,
